CATO-4615 Moved and added some boxes for micro entity.
