+++ src/main/scala/uk/gov/hmrc/ct/accounts/frsse2008/boxes/micro/AC405.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.ct.accounts.frsse2008.micro
+package uk.gov.hmrc.ct.accounts.frsse2008.boxes.micro
 
 import uk.gov.hmrc.ct.accounts.retriever.AccountsBoxRetriever
 import uk.gov.hmrc.ct.box._
 
-case class AC425(value: Option[Int]) extends CtBoxIdentifier(name = "Current Other charges")
+case class AC405(value: Option[Int]) extends CtBoxIdentifier(name = "Current Other Income")
                                      with CtOptionalInteger with Input
                                      with SelfValidatableBox[AccountsBoxRetriever, Option[Int]] {
   override def validate(boxRetriever: AccountsBoxRetriever): Set[CtValidation] = {
