+++ src/main/scala/uk/gov/hmrc/ct/accounts/frsse2008/boxes/micro/AC411.scala
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.ct.accounts.frsse2008.micro
+package uk.gov.hmrc.ct.accounts.frsse2008.boxes.micro
 
 import uk.gov.hmrc.ct.accounts.retriever.AccountsBoxRetriever
 import uk.gov.hmrc.ct.box._
 
-case class AC415(value: Option[Int]) extends CtBoxIdentifier(name = "Current Staff costs")
+
+case class AC411(value: Option[Int]) extends CtBoxIdentifier(name = "Previous Cost of raw materials and consumables")
                                      with CtOptionalInteger with Input
                                      with SelfValidatableBox[AccountsBoxRetriever, Option[Int]] {
   override def validate(boxRetriever: AccountsBoxRetriever): Set[CtValidation] = {
