+++ src/main/scala/uk/gov/hmrc/ct/accounts/frsse2008/boxes/micro/AC415.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.ct.accounts.frsse2008.micro
+package uk.gov.hmrc.ct.accounts.frsse2008.boxes.micro
 
 import uk.gov.hmrc.ct.accounts.retriever.AccountsBoxRetriever
 import uk.gov.hmrc.ct.box._
 
-case class AC426(value: Option[Int]) extends CtBoxIdentifier(name = "Previous Other charges")
+case class AC415(value: Option[Int]) extends CtBoxIdentifier(name = "Current Staff costs")
                                      with CtOptionalInteger with Input
                                      with SelfValidatableBox[AccountsBoxRetriever, Option[Int]] {
   override def validate(boxRetriever: AccountsBoxRetriever): Set[CtValidation] = {
