+++ src/main/scala/uk/gov/hmrc/ct/accounts/frsse2008/boxes/micro/AC416.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.ct.accounts.frsse2008.micro
+package uk.gov.hmrc.ct.accounts.frsse2008.boxes.micro
 
 import uk.gov.hmrc.ct.accounts.retriever.AccountsBoxRetriever
 import uk.gov.hmrc.ct.box._
 
-case class AC410(value: Option[Int]) extends CtBoxIdentifier(name = "Current Cost of raw materials and consumables")
+case class AC416(value: Option[Int]) extends CtBoxIdentifier(name = "Previous Staff costs")
                                      with CtOptionalInteger with Input
                                      with SelfValidatableBox[AccountsBoxRetriever, Option[Int]] {
   override def validate(boxRetriever: AccountsBoxRetriever): Set[CtValidation] = {
