+++ src/main/scala/uk/gov/hmrc/ct/accounts/frsse2008/boxes/micro/AC420.scala
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2017 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.ct.accounts.frsse2008.boxes.micro
+
+import uk.gov.hmrc.ct.accounts.retriever.AccountsBoxRetriever
+import uk.gov.hmrc.ct.box._
+
+case class AC420(value: Option[Int]) extends CtBoxIdentifier(name = "Current Depreciation and other amounts written off assets")
+                                     with CtOptionalInteger with Input
+                                     with SelfValidatableBox[AccountsBoxRetriever, Option[Int]] {
+  override def validate(boxRetriever: AccountsBoxRetriever): Set[CtValidation] = {
+    validateMoney(value)
+  }
+}
