+++ src/main/scala/uk/gov/hmrc/ct/accounts/frsse2008/boxes/micro/AC425.scala
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.ct.accounts.frsse2008.micro
+package uk.gov.hmrc.ct.accounts.frsse2008.boxes.micro
 
 import uk.gov.hmrc.ct.accounts.retriever.AccountsBoxRetriever
 import uk.gov.hmrc.ct.box._
 
-
-case class AC411(value: Option[Int]) extends CtBoxIdentifier(name = "Previous Cost of raw materials and consumables")
+case class AC425(value: Option[Int]) extends CtBoxIdentifier(name = "Current Other charges")
                                      with CtOptionalInteger with Input
                                      with SelfValidatableBox[AccountsBoxRetriever, Option[Int]] {
   override def validate(boxRetriever: AccountsBoxRetriever): Set[CtValidation] = {
