+++ src/main/scala/uk/gov/hmrc/ct/accounts/frsse2008/boxes/micro/AC426.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.ct.accounts.frsse2008.micro
+package uk.gov.hmrc.ct.accounts.frsse2008.boxes.micro
 
 import uk.gov.hmrc.ct.accounts.retriever.AccountsBoxRetriever
 import uk.gov.hmrc.ct.box._
 
-case class AC406(value: Option[Int]) extends CtBoxIdentifier(name = "Previous Other Income")
+case class AC426(value: Option[Int]) extends CtBoxIdentifier(name = "Previous Other charges")
                                      with CtOptionalInteger with Input
                                      with SelfValidatableBox[AccountsBoxRetriever, Option[Int]] {
   override def validate(boxRetriever: AccountsBoxRetriever): Set[CtValidation] = {
