+++ src/main/scala/uk/gov/hmrc/ct/accounts/frsse2008/boxes/micro/AC492.scala
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2017 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.ct.accounts.frsse2008.boxes.micro
+
+import uk.gov.hmrc.ct.box.{CtBoxIdentifier, CtOptionalBoolean, Input}
+
+case class AC492(value: Option[Boolean]) extends CtBoxIdentifier(name = "Micro note? Advances and credits")
+  with CtOptionalBoolean
+  with Input
