+++ src/main/scala/uk/gov/hmrc/ct/accounts/frsse2008/calculations/ProfitOrLossCalculator.scala
@@ -18,7 +18,7 @@ package uk.gov.hmrc.ct.accounts.frsse2008.calculations
 
 import uk.gov.hmrc.ct.accounts.AC12
 import uk.gov.hmrc.ct.accounts.frsse2008._
-import uk.gov.hmrc.ct.accounts.frsse2008.micro._
+import uk.gov.hmrc.ct.accounts.frsse2008.boxes.micro._
 import uk.gov.hmrc.ct.box.CtOptionalInteger
 import uk.gov.hmrc.ct.{MicroEntityFiling, StatutoryAccountsFiling}
