+++ src/main/scala/uk/gov/hmrc/ct/accounts/frsse2008/formats/package.scala
@@ -19,7 +19,7 @@ package uk.gov.hmrc.ct.accounts.frsse2008
 import play.api.libs.json.Format
 import uk.gov.hmrc.ct.accounts.AC12
 import uk.gov.hmrc.ct.accounts.frsse2008.boxes._
-import uk.gov.hmrc.ct.accounts.frsse2008.micro._
+import uk.gov.hmrc.ct.accounts.frsse2008.boxes.micro._
 import uk.gov.hmrc.ct.box.formats.{OptionalBooleanFormat, OptionalDateFormat, OptionalIntegerFormat}
 
 package object formats {
