+++ src/main/scala/uk/gov/hmrc/ct/accounts/frsse2008/retriever/Frsse2008AccountsBoxRetriever.scala
@@ -19,7 +19,7 @@ package uk.gov.hmrc.ct.accounts.frsse2008.retriever
 import uk.gov.hmrc.ct.accounts.AC12
 import uk.gov.hmrc.ct.accounts.frsse2008._
 import uk.gov.hmrc.ct.accounts.frsse2008.boxes._
-import uk.gov.hmrc.ct.accounts.frsse2008.micro._
+import uk.gov.hmrc.ct.accounts.frsse2008.boxes.micro._
 import uk.gov.hmrc.ct.accounts.retriever.AccountsBoxRetriever
 import uk.gov.hmrc.ct.box.retriever.FilingAttributesBoxValueRetriever
