+++ src/test/scala/uk/gov/hmrc/ct/accounts/frsse2008/stubs/StubbedAccountsBoxRetriever.scala
@@ -19,7 +19,7 @@ package uk.gov.hmrc.ct.accounts.frsse2008.stubs
 import uk.gov.hmrc.ct.accounts._
 import uk.gov.hmrc.ct.accounts.frsse2008._
 import uk.gov.hmrc.ct.accounts.frsse2008.boxes._
-import uk.gov.hmrc.ct.accounts.frsse2008.micro._
+import uk.gov.hmrc.ct.accounts.frsse2008.boxes.micro._
 import uk.gov.hmrc.ct.accounts.frsse2008.retriever.Frsse2008AccountsBoxRetriever
 import uk.gov.hmrc.ct.box.CtValue
 import uk.gov.hmrc.ct.box.retriever.FilingAttributesBoxValueRetriever
